Whack-a-mole^Wbug.
A whole load of  bug fixes to my useless attempt at guessing how to do stuff.
I'm actually learning Java from this, which is perhaps not a bad thing.  After a
few hours headbutting Maven and fixing compilation errors, I finally managed to
build a .jar of the plugin.  Now, to test \o/