Refactor out nick-coloring & include groups
If we change the config of Essentials Chat to allow the colours we set to be
applied, it disables the color prefixes which Group Manager applies.

So, to be able to color nicks at all, we need to handle the group-based
colouring, too.  So, hook in to Group Manager to find out the user's group, and
set their colour based on that, if their colour isn't overriden by being a
newbie or being protected.