Disable eating chorus fruit.
Chorus fruit can teleport you at random.  Disable consumption of it for now.

A better fix may be to detect the actual teleportation, check the reason for it,
and if it's because of chorus fruit, and the destination is in someone else's
claim, cancel it.

Howver, the Spigot API doesn't seem to know about chorus fruit as a TP reason,
so we'd have to check for "UNKNOWN" as the reason in the TeleportCause Enum,
which could possibly accidentally match other stuff that shouldn't be blocked.

This will do for now :)