Log block placements
Temporary, to see if I can figure out why we're having block placement issues.

If we see the BlockPlaceEvent events firing, but the blocks don't get placed, it
rules out protocol level things and suggests something might be cancelling the
events, or MC itself is just refusing to accept them.  If we don't see them
fire, then something more low level is wrong.

On the downside, it'll made the console log quite spammy, but at least it's only
when /wedebug is enabled.