Catch and log book edits.
We want to be able to identify who wrote offensive books - as it stands, this
appears not possible unless the book has been signed, which someone leaving an
offensive message naturally won't do.

TODO: see if we can write the author into the NBT data even for an unsigned
book, so if you have the book to hand you can inspect it.