Skip PvP protetection if attacker is also victim.
Thanks to Okyuria for reporting this one.

The immediate case spotted was enderpearl teleportation damage causing
the PvP protection message to appear.

Fixed more generally by checking that attacker is victim, rather than
being specific and recognising if the Projectile is an EnderPearl, so
that it also covers other cases - like managing to hit yourself with
your own arrow/trident etc.