Teleport player on exiting autominecart.
When exiting an auto minecart, players can fall through a one block high
platform.

This used to occasionally happen to people (usually on one block high
platforms in the End above the void!), but since 1.17, it's reliably
reproducible.

It seems that removing the cart leaves the player sitting at least half
way in to the block the rail is on, and they then fall through.

I tried adding a call to eject() on the cart before removing it, which
you'd think would trigger the usual MC logic to find a suitable place to
put the player, but that didn't work.

So, get the player's location, remove the minecart, then teleport them
to the same location but a block higher.

Fixes #14